refactor: change file name from board to boards
